Different loot candle drops per drop (#78)
* Different loot candle drops per drop

The loot table will be initialized in MonsterInformationProvider init; but this class is set up as a singleton (see the protected constructor, and public getInstance) so the loot table is only created once.

* Different loot candle drops per drop

The loot table will be initialized in MonsterInformationProvider init; but this class is set up as a singleton (see the protected constructor, and public getInstance) so the loot table is only created once.